Added exception handling and data validation

+++ src/main/java/com/example/taskmanager/service/TaskServiceImpl.java
@@ -1,7 +1,11 @@
 package com.example.taskmanager.service;
 
 import com.example.taskmanager.entity.Task;
+import com.example.taskmanager.exception.ErrorMessages;
+import com.example.taskmanager.exception.InvalidTaskException;
+import com.example.taskmanager.exception.TaskNotFoundException;
 import com.example.taskmanager.repository.TaskRepository;
+import com.example.taskmanager.validation.TaskValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -15,6 +19,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class TaskServiceImpl implements TaskService {
     private final TaskRepository taskRepository;
+
     @Override
     public List<Task> getTasks() {
         return taskRepository.findAll();
@@ -22,21 +27,34 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public void createTask(Task newTask) {
+        TaskValidator.validateTask(newTask);
         taskRepository.save(newTask);
     }
 
     @Override
     public void updateTask(Long id, Task updateTask) {
+        if(!taskRepository.existsById(id)) {
+            throw new TaskNotFoundException(String.format(ErrorMessages.TASK_NOT_FOUND, id));
+        }
+        TaskValidator.validateTask(updateTask);
+        if (!id.equals(updateTask.getId())) {
+            throw new IllegalArgumentException(ErrorMessages.ID_MISMATCH);
+        }
         taskRepository.save(updateTask);
     }
 
     @Override
-    public Optional<Task> getTask(Long id) {
-        return taskRepository.findById(id);
+    public Task getTask(Long id) {
+        return taskRepository.findById(id)
+                .orElseThrow(() -> new TaskNotFoundException(String.format(ErrorMessages.TASK_NOT_FOUND, id)));
     }
 
     @Override
     public void removeTask(Long id) {
+        if(!taskRepository.existsById(id)) {
+            throw new TaskNotFoundException(String.format(ErrorMessages.TASK_NOT_FOUND, id));
+        }
         taskRepository.deleteById(id);
     }
+
 }
